dom/plugins/test/mochitest/test_crash_notify_no_report.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<window title="Basic Plugin Tests"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script type="application/javascript" src="plugin-utils.js"></script>
  <script type="application/javascript">
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);
  </script>
<body xmlns="http://www.w3.org/1999/xhtml" onload="runTests()">
<embed id="plugin1" type="application/x-test" width="200" height="200"></embed>
</body>
<script class="testbody" type="application/javascript">
<![CDATA[
const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

SimpleTest.waitForExplicitFinish();

var success = false;

var observerFired = false;

var testObserver = {
  observe: function(subject, topic, data) {
    observerFired = true;
    ok(true, "Observer fired");
    is(topic, "plugin-crashed", "Checking correct topic");
    is(data,  null, "Checking null data");
    ok((subject instanceof Ci.nsIPropertyBag2), "got Propbag");
    ok((subject instanceof Ci.nsIWritablePropertyBag2),
"got writable Propbag");

    var id = subject.getPropertyAsAString("pluginDumpID");
    isnot(id, "", "got a non-empty crash id");
    ok(subject.hasKey("additionalMinidumps"), "the additionalMinidumps key is present");
    let directoryService =
      Cc["@mozilla.org/file/directory_service;1"].
      getService(Ci.nsIProperties);
    let pendingD = directoryService.get("UAppData",
                                        Ci.nsIFile);
    pendingD.append("Crash Reports");
    pendingD.append("pending");
    let dumpFile = pendingD.clone();
    dumpFile.append(id + ".dmp");
    ok(dumpFile.exists(), "minidump exists");
    let extraFile = pendingD.clone();
    extraFile.append(id + ".extra");
    ok(extraFile.exists(), "extra file exists");
    Services.crashmanager.ensureCrashIsPresent(id).then(() => {
      // cleanup, to be nice, this will execute before SimpleTest.finish()
      dumpFile.remove(false);
      extraFile.remove(false);
    });
  },

  QueryInterface: function(iid) {
    if (iid.equals(Ci.nsIObserver) ||
        iid.equals(Ci.nsISupportsWeakReference) ||
        iid.equals(Ci.nsISupports))
      return this;
    throw Components.Exception("", Cr.NS_NOINTERFACE);
  }
};


function onPluginCrashed(aEvent) {
  ok(true, "Plugin crashed notification received");
  ok(observerFired, "Observer should have fired first");
  is(aEvent.type, "PluginCrashed", "event is correct type");

  var pluginElement = document.getElementById("plugin1");
  is (pluginElement, aEvent.target, "Plugin crashed event target is plugin element");

  ok(aEvent instanceof PluginCrashedEvent,
     "plugin crashed event has the right interface");

  is(typeof aEvent.pluginName, "string", "pluginName is correct type");
  is(aEvent.pluginName, "Test Plug-in");
  // The app itself may or may not have decided to submit the report, so
  // allow either true or false here.
  ok("submittedCrashReport" in aEvent, "submittedCrashReport is a property of event");
  is(typeof aEvent.submittedCrashReport, "boolean", "submittedCrashReport is correct type");

  var os = Cc["@mozilla.org/observer-service;1"].
           getService(Ci.nsIObserverService);
  os.removeObserver(testObserver, "plugin-crashed");

  // re-set MOZ_CRASHREPORTER_NO_REPORT
  let env = Cc["@mozilla.org/process/environment;1"]
              .getService(Ci.nsIEnvironment);
  env.set("MOZ_CRASHREPORTER_NO_REPORT", "1");
  Services.crashmanager.ensureCrashIsPresent(aEvent.pluginDumpID).then(() => {
    SimpleTest.finish();
  });
}

async function runTests() {
  // the test harness will have set MOZ_CRASHREPORTER_NO_REPORT,
  // ensure that we can change the setting and have our minidumps
  // wind up in Crash Reports/pending
  let env = Cc["@mozilla.org/process/environment;1"]
              .getService(Ci.nsIEnvironment);
  env.set("MOZ_CRASHREPORTER_NO_REPORT", "");

  var os = Cc["@mozilla.org/observer-service;1"].
           getService(Ci.nsIObserverService);
  os.addObserver(testObserver, "plugin-crashed", true);

  document.addEventListener("PluginCrashed", onPluginCrashed, false);

  var pluginElement = document.getElementById("plugin1");
  await SimpleTest.promiseWaitForCondition(() => pluginElement.crash !== undefined,
                                           "Waited too long for plugin to activate");

  try {
    pluginElement.crash();
  } catch (e) {
  }
}
]]>
</script>
</window>
